requirements.txt
selenium
pymongo
flask
webdriver-manager
requests
python-dotenv
